Compilation under VS works Version retriveal fix for Windows Some minor warning fixes DLL implementation for VS / VC++

+++ src/cpu/frequency/windows.cpp
@@ -21,7 +21,7 @@
 unsigned int iware::cpu::frequency() noexcept {
 	LARGE_INTEGER freq;
 	QueryPerformanceFrequency(&freq);
-	return freq.QuadPart * 1000;
+	return static_cast<unsigned int>(freq.QuadPart * 1000);
 }
